+++ app/scripts/routers/base.js
@@ -1,7 +1,7 @@
 import 'backbone-route-filter/backbone-route-filter';
 import Marionette from 'backbone.marionette';
 import Session from 'scripts/facades/session';
-import Routes from 'scripts/helpers/routes';
+import routes from 'scripts/helpers/routes';
 import { props } from 'scripts/decorators';
 
 @props({
@@ -12,7 +12,7 @@ import { props } from 'scripts/decorators';
 export default class BaseRouter extends Marionette.AppRouter {
   redirectIfNotLoggedIn() {
     if (!Session.isLoggedIn()) {
-      this.navigate(Routes.rootPath(), { trigger: true });
+      this.navigate(routes.rootPath(), { trigger: true });
       return false;
     }
   }